Change quizId to setId
As the getSetTitle function is async, a promise is shown in the UI. This
should be fixed in the future. Furthermore, the sorting doesn't make any
sense now, as the setId is not sorted. However, as no timestamps are
logged in the score db, no method to determine the newest/recent games
is present.